Add key for pitch text style, rearrange some code
The key allows configuring the style of the pitch name when displayed
below the string. It is initially set to {font=\tiny} to hopefully avoid
overlapping pitches with accidentals, even with a relatively wide font.

The drawing code now also checks whether the user supplied any fingering
text for below the string, to avoid adding empty space.